requirements.txt
requests
arxiv
pyyaml
google-genai
pydantic